Reorg the kernel
