+++ modelopt/torch/kernels/sparsity/attention/__init__.py
@@ -18,7 +18,11 @@
 import contextlib
 import threading
 
-from modelopt.torch.kernels.common import IS_AVAILABLE, attention, register_triton_attention
+from modelopt.torch.kernels.common.attention import (
+    IS_AVAILABLE,
+    attention,
+    register_triton_attention,
+)
 
 # ---------------------------------------------------------------------------
 # Optional backend registrations (depend on diffusers / ltx_core)
